services/gig-service/requirements.txt
aiohappyeyeballs==2.6.1
aiofiles==23.2.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.12.0
annotated-types==0.7.0
anyio==4.10.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
bitarray==3.7.1
blinker==1.9.0
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
ckzg==2.1.2
click==8.2.1
cloudpickle==3.1.1
colorlog==6.9.0
comm==0.2.3
contourpy==1.3.3
coverage==7.10.6
cryptography==45.0.7
cycler==0.12.1
cytoolz==1.0.1
debugpy==1.8.16
decorator==5.2.1
distlib==0.4.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
executing==2.2.0
fastapi==0.116.1
fastapi-cli==0.0.7
filelock==3.15.4
firebase_admin==7.1.0
Flask==3.0.3
fonttools==4.53.0
frozenlist==1.6.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.3.1
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphviz==0.20.3
greenlet==3.2.4
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.14.0
h2==4.3.0
hexbytes==1.3.1
hpack==4.1.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
imageio==2.35.1
imbalanced-learn==0.6.1
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.32.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.5
lazy_loader==0.4
llvmlite==0.44.0
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mongoengine==0.29.1
msgpack==1.1.1
multidict==6.4.4
narwhals==1.28.0
nest-asyncio==1.6.0
networkx==3.3
numba==0.61.0
numpy==1.26.4
opencv-python==4.10.0.84
optuna==4.2.1
packaging==24.1
pandas==2.2.3
parsimonious==0.10.0
parso==0.8.4
passlib==1.7.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
plotly==6.0.0
pluggy==1.6.0
prompt_toolkit==3.0.50
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.32.1
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.13.0
pyparsing==3.1.2
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2024.1
pyunormalize==16.0.0
PyYAML==6.0.2
pyzmq==26.2.1
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rich-toolkit==0.14.1
rlp==4.1.0
rsa==4.9.1
scikit-image==0.24.0
scikit-learn==1.5.0
scikit-plot==0.3.7
scipy==1.13.1
seaborn==0.9.0
setuptools==70.0.0
shap==0.46.0
shellingham==1.5.4
six==1.16.0
slicer==0.0.8
sniffio==1.3.1
SQLAlchemy==2.0.43
stack-data==0.6.3
starlette==0.47.2
tabulate==0.9.0
threadpoolctl==3.5.0
tifffile==2024.8.30
toolz==1.0.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.2
types-requests==2.32.0.20250515
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2024.1
uliweb-alembic==0.6.9
urllib3==2.4.0
uvicorn==0.35.0
uvloop==0.21.0
virtualenv==20.26.3
watchfiles==1.1.0
wcwidth==0.2.13
web3==7.11.1
websockets==15.0.1
Werkzeug==3.0.3
wheel==0.43.0
yarl==1.20.0
